Tidying up some stuff.
Restructured a bit and changed how the scipt gets the .txt file to work on. There is now a prompt to get the job number, instead of adding in an argument when you start the scipt. It also now checks if the .txt file exists before it continues, if it doesn't it will exit.